requirements.txt
flask
pandas
scikit-learn
joblib
nltk
requests
python-dotenv